+++ cancel.go
@@ -1,0 +1,12 @@
+package httpmux
+
+import (
+	"context"
+	"net/http"
+)
+
+func Cancel(r *http.Request) {
+	_, cancel := context.WithCancel(r.Context())
+
+	cancel()
+}