+++ util.go
@@ -1,0 +1,9 @@
+package httpmux
+
+func resolvedPath(p string) string {
+	if p == "/" {
+		return ""
+	}
+
+	return p
+}